Experimental: h_errno made thread-safe
git-svn-id: https://svn.aros.org/svn/aros/trunk@37460 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/network/stacks/AROSTCP/netlib/writev.c
@@ -31,8 +31,7 @@
 #include <unistd.h>
 #include <sys/uio.h>
 
-int
-writev (int fd, const struct iovec *iov, int iovcnt)
+ssize_t writev (int fd, const struct iovec *iov, int iovcnt)
 {
   int written = 0;
   int res;
